Deploy the opensearch-operator to the opensearch-test namespace
Currently, we are required to deploy an instance of the
opensearch-operator to each namespace where we wish to use it to manage
opensearch clusters.

This commit adds a new release to the opensearch-operator base, so that
the operator is deployed to the opensearch-test namespace.

We add a values-common.yaml file that will be used to configure all
instances of the operator, as well as a values-test.yaml which
configures our test release.

Bug: T404907
Change-Id: Iffe8927992bbd036fe64c56e4e3da0a13e14d85b