Change the name of the cluster deployed in the opensearch-test namespace
We had a clash of release names with both the operator and the cluster
being named 'test' in the opensearch-namespace. The result was that when
deploying the cluster, helmfile removed the operator.

This commit updates the name of the opensearch cluster in the namespace,
so that they do not clash.

Bug: T404907
Change-Id: I7192450760146aa6e4b378e75709191d223342f4