opensearch-operator: add the ability for the operator to watch several namespaces
Now that
https://gitlab.wikimedia.org/repos/data-engineering/opensearch-operator/-/merge_requests/4
was merged and the associated image was published, we can tweak the
opensearch-operator chart to make it watch several namespaces.

We no longer assume that the operator is running in the same namespace
as the cluster, and we make it run into `opensearch-operator` instead.
We thus have to iterate over the list of watched namespaces, and define
the Role in the cluster ns, as well as the RoleBinding, which binds the
Role to the opensearch-operator service account, in the
`opensearch-operator` ns.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T404874
Change-Id: I2be5df9078079fe73bd17f347cc8428f99c64f59