api-gateway: Use metadata to flip csp header handling
We currently completely disable the lua script hook if we want to skip
csp header handling. This disables rate limiting as well.

This change uses the configuration value to set wmf.rest_gateway.no_csp
metadata key, which is then fetched and used to return early in the lua
hook script.

Bug: T406490
Change-Id: I35f97b91a9aa01cbfb1f53217034be91c2850ef3