api-gateway: csp header handling
Use response header append_action ADD_IF_ABSENT instead of lua header
handling.
This disables the lua hooks completely in the absence of rate limiting,
as in that case they were only used to do the csp header manipulation.

Test thoroughly in staging before deploying to production, as this used
to not work.

Bug: T406490
Change-Id: I17e7dc9dbadae600dda0a3cf3f4d3a73b1b3ff02