admin_ng (dse-k8s): watch more OpenSearch-related namespaces
Update the OpenSearch operator config for dse-k8s so it
listens to more namespaces. The new namespaces will be
used for net-new OpenSearch clusters.

We previously applied this in
I79996fbd7b98a82e9cb31f4a52c5b73b5067645e , but we
reverted, thinking the operator was broken. In fact,
the operator image was incorrect. So let's push
this again and test the image with the new feature.

Bug: T357753
Bug: T408012
Change-Id: Ie1a89c25ae87520fdccfc2501e6e06f48e5c5ec4