mw-(api-ext|web): scale next releases to 20% of main
Based on current trends, and consistent with observations during the
8.1 migration, enrollable clients seem to represent 30 - 40% of total
traffic.

In advance of moving to 25% and later 50% enrollment this week, this
scales the next releases of mw-api-ext and mw-web to 20% of main's size,
which we anticipate should be sufficient.

Capacity-wise, we should be able to accommodate this without also
scaling down main in either deployment, due to capacity reclaimed in
I848254c26be994caeaa4b8746dd50440eacf8472.

Bug: T405955
Change-Id: Ib84dde47e98eb778f099080a5adde045d1843410