Deploy airflow images from airflow-dags repository build
We have been using the airflow build process documented on
https://wikitech.wikimedia.org/wiki/Data_Platform/Systems/Airflow/Upgrading to build our base images
for the airflow instances for a while now. However, with the airflow-dags repo refactor with poetry
we wish to test the functionality of the new images built with this method
as we look into future operations.
For this test we shall build the image with the current repository and deploy on airflow-test-k8s
to check the functionality of the image and the overall airflow functions.

Bug: T408711
Change-Id: I9004f58620761240c774a64a374214829845e19a