opensearch-cluster: temporarily remove prometheus-related annotations from chart
I've commented out these annotations,
as they seem to be clobbered rather than merged with higher-priority values files.
We probably won't fix this before we migrate to a newer chart,
so let's disable them and rely on the values files completely for the time being.

Also, add the `prometheus.io/scheme: "https"` annotation in values files, so
prometheus knows to scrape via https instead of http.

Bug: T362114
Bug: T408012
Change-Id: I69796665c3935df0146be3a602516de07ba7b974