mw-web: increase capacity of next
This increases the size of the next release to 50% of the nominal
multi-DC size of main (340 replicas) - i.e., a 1.25x bump over its
previous size.

With 8.3 cookie enrollment at 100%, we're seeing some interesting
effects of having different populations of clients on next vs. main.

One result seems to be that large bumps of cache misses following edits
to widely used templates or modules seem to heavily favour traffic that
is routed to next - i.e., forcing the resulting re-parsing work to
happen there.

In any case, to accommodate this, we should give mw-web yet more
capacity headroom.

Bug: T405955
Change-Id: I85c2fff01f832b7b71c28db4736973f2c86c4bb5