mw-web: Remove the hard-coded k8s version
Why:
Hard-coding kubernetesVersion in helmfiles is a loaded footgun that
we're bound to forget about one day. It is also annoying to update
it everywhere.

What:
Replace it with loading kubernetesVersion from a separate file
that only contains that and is therefore safe to load first.

Also update CI to deal with this correctly.

Note that this commit only fixes a few services, not all of them.
Its purpose is mainly the CI update.

Bug: T388969
Bug: T388390
Depends-On: If2adfbc05656742f9e3ab3a1b158fd3a47fa475d
Depends-On: If155d167ae2cf3524d400e4816d9ca3b87ab3183
Change-Id: Id0b0492ff8eebac7ff5c486384a394c8d2b85c53