Upgrade the superset-production-memcached image to Trixie
We currently do not have a memcached image later than Buster and are currently running
memcached 1.6.6 from a custom backport on Buster.
This upgrades us to trixie available on memcached:1.6.38-1, the changes should be minimal.
The version should work the same without requiring any other changes. We shall first deploy
on superset-test-k8s and then move to production once we confirm everything is working as expected.

Bug: T409151
Change-Id: I258dac039c81f5981834519b519c423c82599be7