Deploy airflow images from airflow-dags repository build
We have been using the airflow build process documented on
https://wikitech.wikimedia.org/wiki/Data_Platform/Systems/Airflow/Upgrading to build our base images
for the airflow instances for a while now. However, with the airflow-dags repo refactor with poetry
and the EOL bullseye-backports we wish to test the functionality of the new images built with this
method and running on bullsye-wikimedia as we look into future operations.
For this test we shall build the image with the current repository where we also recently
upgraded flask-appbuilder to solve some of the initial challenges we had accessing the
'connection/list/' and 'variable/list/' pages. The image works well on devenv
thus we are moving to deploy on and test on airflow-test-k8s

Bug: T408711
Bug: T409271
Change-Id: I3600b9497a75b4b8bceb27ceff03e7452faca697