airflow: define a network policy specific to task pods requiring egress to a proxy
As things were configured, _all_ task pods would be able to egress to the
urldownloader proxy hosts when the `urldownloader` external service was
added to an instance config, such as airflow-platform-eng.

To keep things as tight as possible, we craft a more fine-grained rule:
only task pods labeled with `proxy=urldownloader` running in instances
configured with `worker.proxy.urldownloader.enabled=true` will be able
to egress to the proxy hosts. This way, we don't silently grant possible
internet access to all tasks running on that airflow instance.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T408238
Change-Id: I301e4a9927d3a2ddf772ef716ea1b0ffcc49cdc4