Add a local_upstream_port parameter to the mesh.configuration module
This commit adds a new value to the mesh.configuration module, which
allows the envoy configuration in the tls termination sidecar to specify
a different local port to use as its upstream connection in the local
cluster.

The purpose of this is to support the use cases where an additional
sidecar containers might be required to act as reverse proxy components.

The immediate requirement is for an oauth2-proxy container, which sits
behind envoyproxy.

Bug: T409183
Change-Id: I19c02e873a9c61931b1dbaddb986c9fadc6ddb3e