mw-(api-ext|web): tune maxUnavailable and maxSurge for main
Now that a sizable majority or MediaWiki pods are running 8.3 images,
we now have the opposite of the problem we typically experience at the
start of a migration: cache reuse of 8.1 images is now poor.

While that's tolerable in the case of a typical incremental image build,
it's rather painful in the case of a full image build (e.g., after l10n
updates).

What this does is carry over the same rollingUpdate strategy knobs we
typically use on next, in order to help speed things up - allow a bit
more unavailable pods than usual (6 -> 10%) and yet more surge (3 ->
10%).

We can revert this next week when the migration of these services
completes.

Note that main is intentionally over-provisioned at the moment, so the
bump in maxUnavailable should be fine. Further, we have plenty of spare
headroom for the added surge.

See I962163d84d014abfb29d3d4b471cd25d5ca5a43b and
I99267b3178106dbb43d0f73419160b4cc5755087 for prior art related to the
next release.

Bug: T405955
Change-Id: Ie409eca32f01c09413d9c11a26a67628138fda3c