mw-(api-ext|web): serve 50% of residual traffic on PHP 8.3
This now shifts 50% of residual (80% of nominal multi-DC size) main
release capacity to the migration release, thus serving 50% of residual
traffic on PHP 8.3.

Since the migration release now represents a sizable fraction of overall
capacity, this also dials back the more aggressive maxUnavailable we've
kept on the migration release from the outset of the migration, adopting
the more conservative (but still relaxed) rollingUpdate params currently
present on the main release.

This will be applied in a make-before break fashion across the migration
and main releases, with scale-down of the latter split into two rounds
via local edits (i.e., this patch represents the end-state).

Bug: T405955
Change-Id: I054f61935aa64adae9468de01a93b2e3c32a45b9