mw-(api-ext|web): return capacity from migration to main
This is the first in a series of patches to restore production to a
normal state after the main release has migrated in-place to 8.3.

This will be applied in a make-before-break fashion over 3-4 steps via
local deployment-charts edits (while scap is locked).

In addition, since the main releases are now (1) on running on 8.3 and
(2) back to their normal size, this reverts the relaxed rollingUpdate
params that were added to compensate for poor image cache reuse. See
Ie409eca32f01c09413d9c11a26a67628138fda3c.

Bug: T405955
Depends-On: I3fcc77e82b4848b6d7ff16c424fad151490271c3
Change-Id: Ic0cd54bdb4d68353057a27ba0b22cd5c2da8a01d