mw-(api-ext|web): return main to nominal multi-DC size
This is a pre-step for directing next-release routed traffic back to the
main release.

Once this is applied, we can safely apply the dependent mw-next-routing
ATS Lua and rest-gateway config changes to route cookie-enrolled traffic
back to main.

We should have plenty of capacity to apply this change without yet
removing capacity from the respective next releases [0], particularly if
we also block scap deployments while doing so (which I intend to do).

[0] This happens later in I827f768e4ea0840634823d6ff4c583178f6218ad.

Bug: T405955
Change-Id: I559caa10490f75135fbd3c33c51128c47e89d015