rest-gateway: set 53s timeout for action API
We have imposed the relatively arbitrary timeout of 15s on the action
API, which is generous for most purposes, but is much lower than the
configued timeout of 50s that the search API allows. Given that we're
imposing an *additional* timeout here rather than changing existing
timeouts, there's not much reason to be concerned about increasing
this timeout globally for the short-term. In the longer term, we might
need to add the ability to distinguish between action API actions so
as to set individual timeouts per route.

Bug: T408223
Bug: T410007
Change-Id: Ib4c49790821d98453b3658dca213acb4988dd374