rest-gateway: implement per-route rate limits
This allows rest routes to override the rate limit policy to apply.
This is done by setting metadata on the route, processing it in Lua,
and then setting a header that represents the policy.

This mechanism is quite flexible and aligned with how we do
ratelimit classes.

This also replaces values-rest-devel.yaml and values-rest-ratelimited.yaml with values-minikube.yaml, which is intended for
use in combination with values-staging.yaml. The idea is to keep local testing aligned with the test regime on the staging cluster.
That way, the changes made to values-staging.yaml indicate what
behavior should be tested.

Bug: T409044
Change-Id: I9157b42ba3c0d9273c3962b9ad968fcfc6671bf2