opensearch-cluster: Add cluster ro perms to 'opensearch' user, increase default num of replicas
Give the 'opensearch' user at least as many permissions as the anonymous
user. This doesn't hurt security, and it keeps us from having to switch
back and forth from using/not using a password login.

Additionally, set the default number of replicas to 2 (up from the default of 1),
as promised in the design doc.

Bug: T408012
Change-Id: I7048a18d7c733e6c9a987880c1e7a56c5eba2c03