mw-*: clean up 8.3 migration rollingUpdate and timeout tweaks
Now that the migration is complete, this cleans up various tweaks to the
helmfile operation timeout (now uniformly back to 10m) and rollingUpdate
params.

The latter have been commented out with illustrative examples, and the
explanatory comment updated to reflect our experience this time around.

Notably, this changes the guidance slightly - still allowing for ~ 20%
in-flight container pulls, spread between maxUnavailable and maxSurge,
which burns a bit more spare capacity to enable a more conservative
maxUnavailable.

Bug: T405955
Change-Id: Idaa57fdf41cf76d862fe5e4279ddd2e96272f88e