Add a new deploy-spark-support clusterrole
This is a new clusterrole that will be applied to the deploy users of
certain namespaces. The role takes the default deploy user role and adds
the necessary rights to create roles and rolebindings.

These rights will subsequently be used by the spark-support chart to
permit airflow serviceaccounts in other namespaces to manage task pods
in each namespace to which is it applied.

This is a pattern that we have previously used with the
mediawiki-dumps-legacy clusterrole and corresponding chart.

Bug: T410017
Change-Id: I461d09af5b0a90c2cb6ef7a944c40e59abde25d0