Enable the deploy-spark-support deploy clusterrole for two test namespaces
We have created a new clusterrole that permits the creation of roles and
rolebindings, in addition to the normal rights of the deploy
clusterrole.

This will be applied to a select few namespaces, where we wish to deploy
a spark-support chart. The purpose is to allow Airflow to launch tasks
into the target namespace.

This commit enables the use of this new clusterrole for the 'spark' and
'analytics-test' namespaces.

Bug: T410017
Change-Id: If13dac1d8140325ad6650c417ecdec992d2c48c7