api-gateway: add lua tests
This also fixes a bug that would cause no rate limits to be applied
in test mode, when a x-wmf-user-id is provided by the client directly.
This is not possible in proudction.

NOTE: This patch introduces a number of temporary hacks that are
immediately removed again in the follow-up, Ie0c56bfd58c631f9.
It is intended as a steping stone that ensures that the Lua logic
is not broken by the refactoring in the next patch.
This patch should only be deployed together with the follow-up that
cleans up the temporary hacks.

Change-Id: Ic763c1a114c3a4a64576731912b705fae8a9daa4