Add an RBAC configuration to allow spark driver pods to function
We are now able to get Airflow to create pods in a spark-enabled
namespace. The next step is create the RBAC rules that permit spark
driver pods to manage the kubernetes resources that they need.

In order to do this, we want to add a new `spark` serviceaccount, which
will be used by `spark-submit`.

The core requirement is to be able to manage pods, but there are some
optional rights, around configmaps, services, and namespace quotas.

These rights are explained to some extent here:
https://spark.apache.org/docs/3.5.7/running-on-kubernetes.html#rbac

Change-Id: I7ac424285a2b1b4b1edd3467a8beda7fcebad18b