Remove unnecessary and/or incorrect hadoop/spark config options
The initial Hadoop and Spark configuration files in this chart were
imported from our Airflow chart, where they had in-turn been imported
frlom our puppet-managed hosts.

Therefore, some of the configuration parameters that had been set in
the core-site.xml, hdfs-site.xml, yarn-site.xml and spark-defaults.conf
files were not relevant to the configuration that we are trying to
achieve in a spark-enabled kubernetes namespace.

This commit removes those parameters that are not relevant.

The hive-site.xml file is an interesting case, because we want to use
this with the spark thriftserver, connecting it to an external,
kerberized hive metastore, but listening on localhost port 10000,
without authentication.

Bug: T410017
Bug: T406833
Change-Id: Idf95d58273e9c5319e536f37c13138550078aa99