services: add maps-next.w.o as FQDN for kartotherian staging
The idea is to create a maps-next.w.o external endpoint, routed
through the CDN, that hits Kartotherian in staging. The stack
is completely separated from the production one, including postgres
and the swift cache.

The best way to implement the above in staging is to enable ingress,
since it will allow us to have more redundancy in case of failures
(the staging's discovery endpoint is just a CNAME to kubestage1003).

Add the ingress config to the kartotherian's chart, and enable
it in staging.

Bug: T409528
Change-Id: I40322b0981a4381f58809850dcbe6e07c48500fd