miscweb: add wikipedia25.org to extra SANs
Add wikipedia25.org to certificates used by Kubernetes
Ingress.

We need both variants, bare domain and with www and
additionally the donate subdomain which we will have to
redirect over to wikimediafoundation.org

This will be needed to host a microsite using this domain
on the existing miscweb-k8s service.

Bug: T408592
Bug: T408168
Change-Id: I683d914598a3d990c4eb89b693ae25636e83e8dc