Bump the size of the PostgreSQL WAL for airflow-main to 30GB
We experienced an issue with postgresql running behind airflow-main
while attempting to carry out database maintenance.

The 15 GB WAL volume filled up, which caused the database server to
crash and a replica was promoted.

This commit causes the size of the WAL volume serving airflow-main to be
doubled in size, to 30 GB.

Bug: T375846
Bug: T412003
Change-Id: I8de3e22275a8192fe31be88c574cfa512ec54a4e